requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
numpy>=1.24.0
nibabel>=5.0.0          # for MRI/CT .nii files
SimpleITK>=2.2.0        # for CT/MRI preprocessing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.11.0
einops>=0.7.0
timm>=0.9.0
pyyaml>=6.0
tqdm>=4.65.0
monai>=1.2.0            # medical image transforms
matplotlib>=3.7.0
tensorboard>=2.14.0
# SAM-2 must be installed separately:
# git clone https://github.com/facebookresearch/segment-anything-2
# pip install -e segment-anything-2/